[level 2] Title: [PCCP 기출문제] 2번 / 퍼즐 게임 챌린지, Time: 53.72 ms, Memory: 120 MB -BaekjoonHub

+++ 프로그래머스/2/340212. ［PCCP 기출문제］ 2번 ／ 퍼즐 게임 챌린지/［PCCP 기출문제］ 2번 ／ 퍼즐 게임 챌린지.java
@@ -13,7 +13,7 @@ class Solution {
                     totalTime += times[i];
                 }
                 else{
-                    totalTime +=  (diffs[i]-level)*(times[i]+times[i-1])+times[i];
+                    totalTime +=  (long)(diffs[i]-level)*(times[i]+times[i-1])+times[i];
                 }
             }
